Initial commit
This version of my personal infrastructure is the first to ever
experience the cold embrace of version control.

IT DOES NOT WORK: I have not yet updated the email server to look in the
right place for certificates. However, I am about to make additions to
the services version those changes. Hence this commit.

Things are already much different in this rework then they are running
on my server. The new setup is intended to be elegant, declarative, and
robust. As such, I have:
1. Split my Docker Compose config into several, domain-oriented files.
2. Removed numerous services I had not used in years.
3. Migrated from a complex nginx setup to Caddy.
4. Stopped building images anywhere on the server.
5. Added a Watchtower image for automatic updates.
6. Merged the db and services directories, which now contain versioned
   config as well as server-only live data.

Concision in this repository is a virtue. It indicates a delegation of
concerns—leaving the aspects of running my infrastructure best managed
by experts, or reliant on consistent updates, to be managed elsewhere.

See you soon! Let's get this up and running.