Add backups image
The custom-built container runs the acts script via cron to trigger
tarsnap. The entire services directory is included in the backups. These
are conducted while the services are still running, and thus could
contain corrupt mid-write data—one day I might set up a more correct
solution but for now I will accept the risk, and make a ton of backups
to reduce the likelihood that they're all bricked.

As it is custom, the image is updated via the unregistry.

Source:
https://github.com/FIGBERT/acts-docker