Update Radicale installation
The image was swapped out, as the previous one wasn't built for ARM
architectures. Ideally we would use the official ghcr.io one (which
exists now, apparently) but it doesn't have major version tags. This
one is built with CI from the official Dockerfile and thus is the next
best option.

Configuration is also now versioned, including a SOPS-ified users file.