Switch maddy to rspamd normal worker
The spam filtering has been silently failing since it was booted up. My
theory, addressed in this commit, is that the proxy worker is expecting
messages in the milter format whereas maddy is sending them in rspamd's
native HTTP format. This commit changes endpoints to one that I believe
is expecting such HTTP messages.

Should that not be the case, maddy does have native support for milter.
Indeed, I'm rather confused about which is better to use. But rspamd's
documentation is a travesty and this is the smallest change I can
imagine to achieve the desired effect so it's the one I'm making first.