Fix Scroll Thingy error fallback
It seems that, as the fallback this is addressing is a refused
connection, the error handling block needs to be outside the
reverse_proxy block. Obscure Caddy trivia.

One alternative to this method could be load balancing, with the second
upstream being a fallback static site. I considered this. However, I
decided against it as it would require running a second container which
is a lot of overhead for a single, static HTML file.